+++ CHANGELOG.md
@@ -34,3 +34,6 @@ All notable changes to Equinox will be documented in this file.
 - handle browser history
 - fix card should default to automatic height
 - improve Liquid Glow active heat/cool side glow with a center-out expanding pulse while preserving the Home Assistant card border theme
+- Add option to only display border glow when hvac is active ( default checked )
+- add editor tabs to hide specific HVAC mode and preset buttons based on the attached climate capabilities
+-